Hall Effect Board Changes
Added a voltage divider to get to 2.5V, as there is not a VCU output which can do that - its toggleable by jumper cable, so if we have a better solution we don't need to make a new board to implement it.
Also moved pins to all be in the same row, since most of our boards are designed like that and it makes the overall board slightly less messy.